Fix warnings in Windows build.
Fix a few integer precision warnings.

Tested: Windows Testbed
Bug: 15290318

+++ liquidfun/Box2D/Box2D/Particle/b2ParticleGroup.h
@@ -402,7 +402,7 @@ inline void b2ParticleGroupDef::SetPosition(float32 x, float32 y)
 
 inline void b2ParticleGroupDef::SetColor(int32 r, int32 g, int32 b, int32 a)
 {
-	color.Set(r, g, b, a);
+	color.Set((uint8)r, (uint8)g, (uint8)b, (uint8)a);
 }
 #endif // LIQUIDFUN_EXTERNAL_LANGUAGE_API
